requirements.txt
pydantic
httpx
openai
fastapi
uvicorn[standard]
pyngrok
python-dotenv
requests
